requirements.txt
# Article Scraper 依赖

# HTTP 请求（curl_cffi 绕过 Cloudflare）
curl_cffi>=0.5.0
requests>=2.31.0

# HTML 解析
beautifulsoup4>=4.12.0
lxml>=4.9.0

# HTML 转 Markdown
markdownify>=0.11.0

# 浏览器自动化（用于反爬网站）
playwright>=1.40.0

# 数据处理
python-dateutil>=2.8.0

# BigQuery 写入
google-cloud-bigquery>=3.0.0

# 可选：代理池支持（如果需要连接外部代理池服务）
# httpx>=0.25.0
